Update version to 0.3.1, add run-examples target to Makefile, and enhance example notebooks with execution metadata. Remove deprecated execution_safer.py file and refactor execution handling in execution.py for improved security and clarity.
