+++ Makefile
@@ -1,4 +1,4 @@
-.PHONY: publish test clean
+.PHONY: publish test clean run-examples
 
 publish: clean test
 	@echo "Building distribution files..."
@@ -10,7 +10,12 @@ test:
 	@echo "Running tests..."
 	pytest
 
+run-examples:
+	@echo "Running example notebooks..."
+	python scripts/run_examples.py --max-workers 3
+
 clean:
 	rm -rf build/
 	rm -rf dist/
 	rm -rf *.egg-info/
+	rm -rf examples/executed/
