+++ examples/README.md
@@ -3,3 +3,9 @@
 Some notebooks with different examples.
 
 Note: Best to view this through NBViewer here (makes plots visible): https://nbviewer.org/github/andrewm4894/plot-agent/tree/main/examples/
+
+To run (in-place) all example notebooks:
+
+```bash
+make run-examples
+```
