+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "plot-agent"
-version = "0.3.0"
+version = "0.3.1"
 authors = [
     { name="andrewm4894", email="andrewm4894@gmail.com" },
 ]
