+++ requirements-dev.txt
@@ -1,0 +1,3 @@
+pytest
+pytest-cov
+papermill
