+++ requirements.txt
@@ -6,5 +6,3 @@ langchain-core
 langchain
 langchain-openai
 pydantic
-pytest
-pytest-cov
